requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
